# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bedrock SSO Proxy API",
    "description": "JWT-authenticated HTTP proxy server that provides secure access to AWS Bedrock APIs",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin User Management"
        ],
        "summary": "List users",
        "description": "Get a paginated list of all users with optional sorting",
        "operationId": "list_users",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of users to return (default: 50, max: 500)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of users to skip for pagination (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order: created_at, updated_at, last_login, email (default: created_at)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort direction: asc or desc (default: desc)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
    "/api/admin/users/search": {
      "get": {
        "tags": [
          "Admin User Management"
        ],
        "summary": "Search users",
        "description": "Search for users by email, display name, or provider user ID",
        "operationId": "search_users",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search term to match against email, display_name, or provider_user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of users to return (default: 50, max: 500)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of users to skip for pagination (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid search parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}": {
      "get": {
        "tags": [
          "Admin User Management"
        ],
        "summary": "Get user by ID",
        "description": "Get detailed information about a specific user",
        "operationId": "get_user_by_id",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}/state": {
      "put": {
        "tags": [
          "Admin User Management"
        ],
        "summary": "Update user state",
        "description": "Update the state of a user (active, disabled, or expired)",
        "operationId": "update_user_state",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User state updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "API Error response for documentation",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "UpdateUserStateRequest": {
        "type": "object",
        "description": "Request model for updating user state",
        "required": [
          "state"
        ],
        "properties": {
          "state": {
            "$ref": "#/components/schemas/UserRecord",
            "description": "New user state"
          }
        }
      },
...
      "UserListResponse": {
        "type": "object",
        "description": "Response model for user list",
        "required": [
          "users",
          "total",
          "count",
          "offset",
          "limit"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of users returned in this response",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "description": "Limit used for pagination",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "description": "Offset used for pagination",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "description": "Total number of users matching the query",
            "minimum": 0
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "description": "List of users"
          }
        }
      },
      "UserRecord": {
        "type": "string",
        "description": "User state enum for tracking account status",
        "enum": [
          "active",
          "disabled",
          "expired"
        ]
      },
      "UserResponse": {
        "type": "object",
        "description": "User response model for API",
        "required": [
          "id",
          "provider_user_id",
          "provider",
          "email",
          "created_at",
          "updated_at",
          "state"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name from OAuth provider"
          },
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "User ID"
          },
          "last_login": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Last login timestamp"
          },
          "provider": {
            "type": "string",
            "description": "OAuth provider name"
          },
          "provider_user_id": {
            "type": "string",
            "description": "Provider user ID (external ID from OAuth provider)"
          },
          "state": {
            "$ref": "#/components/schemas/UserRecord",
            "description": "User account state"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "jwt_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "x_api_key_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}